pyproject.toml
[project]
name = "queai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "arxiv>=2.2.0",
  "fastapi[standard]>=0.116.1",
  "google-genai>=1.29.0",
  "grandalf>=0.8",
  "ipython>=9.4.0",
  "langchain-community>=0.3.27",
  "langchain-experimental>=0.3.4",
  "langchain-tavily>=0.2.10",
  "langchain[google-genai]>=0.3.26",
  "langgraph>=0.5.3",
  "langgraph-checkpoint-sqlite>=2.0.11",
  "pillow>=11.3.0",
  "pymupdf>=1.26.4",
  "wikipedia>=1.4.0",
]
